Move IConfigurationField and IConfigurationModule to lodestar-cli
