+++ packages/lodestar-cli/src/lodestar/db/options.ts
@@ -1,0 +1,17 @@
+import {IConfigurationModule} from "../util/config";
+
+export const DatabaseOptions: IConfigurationModule = {
+  name: "db",
+  fields: [
+    {
+      name: "name",
+      description: "Path to file database",
+      configurable: true,
+      type: String,
+      cli: {
+        flag: "db",
+        short: "d"
+      }
+    }
+  ]
+};
