+++ packages/lodestar-cli/src/lodestar/node/options.ts
@@ -1,0 +1,24 @@
+import {ChainOptions} from "../chain/options";
+import {DatabaseOptions} from "../db/options";
+import {Eth1Options} from "../eth1/options";
+import {NetworkOptions} from "../network/options";
+import {OpPoolOptions} from "../opPool/options";
+import {SyncOptions} from "../sync/options";
+import {BeaconLoggerOptions} from "./loggerOptions";
+import {MetricsOptions} from "../metrics/options";
+import {IConfigurationModule} from "../util/config";
+
+export const BeaconNodeOptions: IConfigurationModule = {
+  name: "config",
+  fields: [
+    ChainOptions,
+    DatabaseOptions,
+    // PublicApiOptions,
+    Eth1Options,
+    NetworkOptions,
+    OpPoolOptions,
+    SyncOptions,
+    BeaconLoggerOptions,
+    MetricsOptions,
+  ]
+};
