+++ packages/lodestar-cli/src/lodestar/opPool/options.ts
@@ -1,0 +1,6 @@
+import {IConfigurationModule} from "../util/config";
+
+export const OpPoolOptions: IConfigurationModule = {
+  name: "opPool",
+  fields: []
+};
