+++ packages/lodestar-cli/src/lodestar/sync/options.ts
@@ -1,0 +1,6 @@
+import {IConfigurationModule} from "../util/config";
+
+export const SyncOptions: IConfigurationModule = {
+  name: "sync",
+  fields: []
+};
