+++ packages/lodestar-cli/src/lodestar/util/config.ts
