+++ packages/lodestar/package.json
@@ -52,7 +52,6 @@
     "@chainsafe/eth2.0-utils": "0.1.0",
     "@chainsafe/lodestar-validator": "0.2.0",
     "@chainsafe/ssz": "0.5.2",
-    "@iarna/toml": "^2.2.3",
     "@types/async": "^3.0.1",
     "assert": "^2.0.0",
     "async": "^3.1.0",
