+++ packages/lodestar/src/api/options.ts
@@ -1,6 +1,5 @@
 import rpcDefaultOptions, {IRpcOptions} from "./rpc/options";
-import restApiDefaultOptions, {IRestApiOptions, RestOptions} from "./rest/options";
-import {IConfigurationModule} from "../util/config";
+import restApiDefaultOptions, {IRestApiOptions} from "./rest/options";
 
 
 export interface IApiOptions {
@@ -11,12 +10,4 @@ export interface IApiOptions {
 export default {
   rpc: rpcDefaultOptions,
   rest: restApiDefaultOptions
-};
-
-export const ApiOptions: IConfigurationModule = {
-  name: "api",
-  description: "Options related to api interfaces",
-  fields: [
-    RestOptions,
-  ]
 };
