+++ packages/lodestar/src/api/rest/options.ts
@@ -1,6 +1,4 @@
 import {ApiNamespace} from "../index";
-import {booleanOption, IConfigurationModule} from "../../util/config";
-import {processApiNamespaces} from "../utils";
 
 export interface IRestApiOptions {
   enabled: boolean;
@@ -16,45 +14,4 @@ export default {
   host: "127.0.0.1",
   port: 9596,
   cors: "*"
-};
-
-export const RestOptions: IConfigurationModule = {
-  name: "rest",
-  description: "Options to configure rest api server",
-  fields: [
-    booleanOption("enabled", "--rest"),
-    {
-      name: "api",
-      type: "string",
-      process: processApiNamespaces,
-      configurable: true,
-      cli: {
-        flag: "--rest-api"
-      }
-    },
-    {
-      name: "host",
-      type: "string",
-      configurable: true,
-      cli: {
-        flag: "--rest-host"
-      }
-    },
-    {
-      name: "port",
-      type: "number",
-      configurable: true,
-      cli: {
-        flag: "--rest-port"
-      }
-    },
-    {
-      name: "cors",
-      type: "string",
-      configurable: true,
-      cli: {
-        flag: "--rest-cors"
-      }
-    }
-  ]
 };
