+++ packages/lodestar/src/eth1/options.ts
@@ -1,6 +1,5 @@
 import {number64} from "@chainsafe/eth2.0-types";
 import {BaseProvider} from "ethers/providers";
-import {IConfigurationModule} from "../util/config";
 import {EventFragment, FunctionFragment, Interface, ParamType, BigNumber} from "ethers/utils";
 
 export interface IEth1Options {
@@ -16,59 +15,6 @@ export interface IEth1Options {
     bytecode: string;
   };
 }
-
-export const Eth1Options: IConfigurationModule = {
-  name: "eth1",
-  description: "Eth 1.x options",
-  fields: [
-    {
-      name: "provider",
-      fields: [
-        {
-          name: "url",
-          description: "Url to eth1 node with enabled rpc api",
-          type: String,
-          configurable: true,
-          cli: {
-            flag: "eth1RpcUrl"
-          }
-        },
-        {
-          name: "network",
-          description: "Eth1 network id",
-          type: "number",
-          configurable: true,
-          cli: {
-            flag: "networkId"
-          }
-        }
-      ]
-    },
-    {
-      name: "depositContract",
-      fields: [
-        {
-          name: "deployedAt",
-          description: "Block number at which contract is deployed",
-          type: "number",
-          configurable: true,
-          cli: {
-            flag: "depositContractBlockNum"
-          }
-        },
-        {
-          name: "address",
-          description: "Address of deposit contract",
-          type: String,
-          configurable: true,
-          cli: {
-            flag: "depositContract"
-          }
-        }
-      ]
-    }
-  ]
-};
 
 const config: IEth1Options = {
   provider: {
