+++ packages/lodestar/src/network/options.ts
@@ -1,5 +1,4 @@
 import {number64} from "@chainsafe/eth2.0-types";
-import {IConfigurationModule} from "../util/config";
 
 export interface INetworkOptions {
   maxPeers: number64;
@@ -9,45 +8,6 @@ export interface INetworkOptions {
   connectTimeout: number;
   disconnectTimeout: number;
 }
-
-export const NetworkOptions: IConfigurationModule = {
-  name: "network",
-  fields: [
-    {
-      name: "maxPeers",
-      type: "number",
-      configurable: true,
-      process: (input) => {
-        return parseInt(input);
-      },
-      cli: {
-        flag: "maxPeers"
-      }
-    },
-    {
-      name: "bootnodes",
-      type: [String],
-      configurable: true,
-      process: (input: string) => {
-        return input.split(",").map((input) => input.trim());
-      },
-      cli: {
-        flag: "bootnodes"
-      }
-    },
-    {
-      name: "multiaddrs",
-      type: [String],
-      configurable: true,
-      process: (input: string) => {
-        return input.split(",").map((input) => input.trim());
-      },
-      cli: {
-        flag: "multiaddrs"
-      }
-    },
-  ]
-};
 
 const config: INetworkOptions = {
   maxPeers: 25,
