+++ packages/lodestar/src/node/options.ts
@@ -2,16 +2,15 @@
  * @module node
  */
 
-import defaultChainOptions, {ChainOptions, IChainOptions} from "../chain/options";
-import defaultDatabaseOptions, {DatabaseOptions, IDatabaseOptions} from "../db/options";
+import defaultChainOptions, {IChainOptions} from "../chain/options";
+import defaultDatabaseOptions, {IDatabaseOptions} from "../db/options";
 import defaultApiOptions, {IApiOptions} from "../api/options";
-import defaultEth1Options, {Eth1Options, IEth1Options} from "../eth1/options";
-import defaultNetworkOptions, {INetworkOptions, NetworkOptions} from "../network/options";
-import defaultOpPoolOptions, {IOpPoolOptions, OpPoolOptions} from "../opPool/options";
-import defaultSyncOptions, {ISyncOptions, SyncOptions} from "../sync/options";
-import defaultLoggerOptions, {BeaconLoggerOptions, IBeaconLoggerOptions} from "./loggerOptions";
-import defaultMetricsOptions, {IMetricsOptions, MetricsOptions} from "../metrics/options";
-import {IConfigurationModule} from "../util/config";
+import defaultEth1Options, {IEth1Options} from "../eth1/options";
+import defaultNetworkOptions, {INetworkOptions} from "../network/options";
+import defaultOpPoolOptions, {IOpPoolOptions} from "../opPool/options";
+import defaultSyncOptions, {ISyncOptions} from "../sync/options";
+import defaultLoggerOptions, {IBeaconLoggerOptions} from "./loggerOptions";
+import defaultMetricsOptions, {IMetricsOptions} from "../metrics/options";
 
 export interface IBeaconNodeOptions {
   chain: IChainOptions;
@@ -24,21 +23,6 @@ export interface IBeaconNodeOptions {
   logger: IBeaconLoggerOptions;
   metrics: IMetricsOptions;
 }
-
-export const BeaconNodeOptions: IConfigurationModule = {
-  name: "config",
-  fields: [
-    ChainOptions,
-    DatabaseOptions,
-    // PublicApiOptions,
-    Eth1Options,
-    NetworkOptions,
-    OpPoolOptions,
-    SyncOptions,
-    BeaconLoggerOptions,
-    MetricsOptions,
-  ]
-};
 
 const config: IBeaconNodeOptions = {
   chain: defaultChainOptions,
