+++ packages/lodestar/src/sync/options.ts
@@ -1,14 +1,7 @@
-import {IConfigurationModule} from "../util/config";
-
 export interface ISyncOptions {
   blockPerChunk: number;
 }
 
-
-export const SyncOptions: IConfigurationModule = {
-  name: "sync",
-  fields: []
-};
 
 const config: ISyncOptions = {
   blockPerChunk: 20
